refactor: reorganize methods
